Начата работа над списком событий. Нужнно откорректировать вывод списка пользователей спикиров или принявших участие в событии

+++ app/src/main/java/ru/maxpek/friendslinkup/entity/EventRemoteKeyEntity.kt
@@ -4,10 +4,10 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class EventRemoteKeyEntity (
+data class EventRemoteKeyEntity(
     @PrimaryKey
     val type: KeyType,
-    val id: Long,
+    val id: Int,
     ) {
     enum class KeyType {
         AFTER, BEFORE
